+++ src/pages/home/contents/DashboardPage/hook/index.ts
@@ -1,1 +1,2 @@
 export { default as useSelectedTagMemosManager } from './useSelectedTagMemosManager';
+export { default as useTagsManager } from './useTagManager';
